Remove X_XF86VidModeGetGammaRampSize checks.
The protocol defines were moved to a different header in recent versions of the
XF86VidMode protocol packages, which breaks the build.

The GammaRamp related entry points were added to the XF86VidMode extension in
version 2.1.  Years ago, it was not uncommon to try to build nvidia-settings
with a pre-2.1 version of xf86vmode.h, but I don't think that happens frequently
any more.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>