Remove the Enable Tooltips option since it doesn't work with GTK 2.12.0 and higher
GTK 2.12 change how tooltips are handled internal to GTK. This addition, while
leaving the function interface, changed the internals and made the old
enable/disable and set_delay functions as No-Ops.

This change removes the configuration option from the config page for all
version of GTK. GTK 3.x only uses the later version of tooltips and is
functioanlly uneffected by this change.