nvidia-settings: Fix crash when failing to initialize sensors
nvidia-settings' sensor code gets pretty confused with >1 sensor per
GPU, as with GF100GL (p1030).  The X driver finds 3 sensors, but the
nvml backend only enumerates 1.  ctk_thermal_new() will query the X
driver to figure out which sensors are on the GPU, then query NVML to
enumerate data from those sensors.  In a dual-p1030 configuration,
nvidia-settings will thus expose both GPU 0's and GPU 1's temperature
data on the page for GPU 0 and nothing for GPU 1.

Here, the ctk_thermal->sensor_info array was being allocated
uninitialized, and when the NvCtrlGetTarget() call failed for a
particular sensor, it wasn't being cleared to NULL.  Later, when we
attempted to use that data, we could crash.

This change converts the malloc to an nvalloc (which also fixes a bug
where if the malloc had failed, we could have crashed assigning to the
array).