PRIME Sync read-only support on Display Configuration page.
This change adds support to nvidia-settings for showing known PRIME Displays on
the Display Configuration page in the layout window and content section.
Selecting the PRIME Display is supported as well as displaying associated info
but it is not possible to move, rotate, or change the displays attributes at all
from nvidia-settings. This is due to our reliance on external configuration
tools to configure the external driver and display.

Note that PRIME without Synchronization is limited in the information we have
available to display. Currently the PRIME display name is the only attribute we
require PRIME Sync for but that is expected to change. The UI has default
behaviors and text for displays that are missing this extra information.