Add HDMI 3D Label to refresh rate on the display config page
When screens are configured to use HDMI 3D stereo, the refresh rate listed as
half of what the mode normally describes. This label helps the user to confirm
the reason the value may be different than expected.

This brings the Display Config page more in link with the Display Device and
Framelock pages that show the refresh rate in a similiar manner with the 'HDMI
3D' label.