Don't repeat identical queries with Display Targets when querying all attributes.
When NV-CONTROL tries to get the Dpy pointer for a queried target and that
target is itself a Display, we currently ignore the mask argument and return the
original Target. The problem occurs when we run the command, "--query all". With
non-display targets, we iterate over a display mask to make sure we query
displays associated with the Screen, GPU, etc and this works fine. If we are
using a Display Target, the mask is ignore and we get 24 matches for each bit in
the mask because we interally ignore the mask. This change exits the bit mask
iterator loop after the first run when we are using a Display Target.