Always check if the ECC Configuration Pending state has changed.
Right now on the ECC status page in nvidia-settings we poll for changes in the ECC state of the GPU. When ECC is enabled we poll for ECC errors. The problem is we only check the Configuration Pending state when the application starts up. We can detect via nvcontrol events, if this value is changed for example on the command line by "nvidia-settings -a ECCConfiguration=1". But if this value is changed by a non-nvcontrol method like "nvidia-smi -e 1", we don't detect or display the change.

This change adds querying this value to the polling loop of the ECC Status page.