Keep the ECC Reset Default Config button in sync with current ECC pending status.
Anytime the ECC Pending Status Checkbox is set to a different value than the ECC Default Configuration value, this button should be enabled. It should be clear to the user that what the Reset button's value should be based on behavior alone.

This change updates the enabled/disabled state of the Reset anytime we receive a update to the pending config value and or when the user changes the value manually.