Fix for nvidia-settings throws dbus error when nvidia-gridd is not running.
Issue:
Getting assertion on lauching nvidia-settings when nvidia-gridd is not running:
"ERROR: GRID License dbus communication: Argument is not int! "

Fix:
The current dbus API queues a message to send(dbus_connection_send_with_reply)
and block until we receive a reply.
Now changed the dbus API(dbus_connection_send_with_reply_and_block)
which sends a message and blocks for a default time period while waiting for a reply.
This dbus function has advantage over the current API as it returns an error code on failure.
And depending on the error code returned we can decide whether to send the signal to nvidia-gridd or not.