Allow moving displays to the origin even if the layout window is too large.
Currently, we can get in a situation where the nvidia-settings display layout window is setup to be larger than what the X server allows. The problem is if we are in this situation, the user may not be able to fix the layout because our enforcement of size limits doesn't allow the user to move displays to the origin. This is because when moving items to the left or upward, we calculate our movement limit with respect to the right-most or furthest down position used.

This change will allow us to always move items to the origin. Note that we already always redefine the origin to be the top-left most valid item. This position should always be available when moving items.