Allow Mosaic configuration when a screen is selected.
Right now, the Enable Mosaic button is only validated properly and shown when a Display object is selected on the Display Config page of nvidia-settings. Mosaic support is a feature of the gpu and shouldn't depend on if a Display or Screen is selected. This change adds the validation and activation of the Mosaic feature with a Screen selection in addition to the existing Display selection.