Fix for un-initialized variables valgrind error
The variable on this change are liable to not be initialized if the
function that is meant to initialize errors out. And the uninitialized
variable would still be passed for the creation of a widget. Since in
the case of this happening such widget won't be used anyways, chances
of this causing trouble are slim, but still we shouldn't have it.
Additionally it shows on valgrind so it creates noise when trying to
troubleshoot something else.