Add text to the help page for widgets missing entries.
This change adds help text for the "Save Current Configuration" button on the "nvidia-settings Configuration" page and re-enables a flag that allows the help text for "Sync to this display device" on the XVideo Settings page.

Also removes the per entry tooltip text for the "Sync to this display device" and adds the tooltip to the entire container widget since it exists for the entire lifetime of the application.