Stereo Test
It had been requested to provide end users with a way to test if stereo mode has
been configured and is working properly.

To address such, it was decided to create a simple demo/test and to get such app
embedded in nvidia settings by creating a custom gtk widget on which to render
opengl. (Such widget seems to exist, but only in versions of gtk3 that are newer
to what nvidia-settings use, hence the need to create our own).

This implementation will work with gtk3 and ignored on gtk2.