Setting feature type correctly on 'Manage GRID License' page creation.
Issue:
When nvidia-settings is launched on a Quadro vDWS Edition licensed VM, the
message shown on 'Manage GRID License' page as well as the one logged in
/var/log/messages incorrectly states that the system has switched to Tesla
(unlicensed) and user needs to reboot the system.

Root cause and fix:
Issue was observed because the feature type was not set correctly on page
creation. Fixed it with this change.

Testing done:
Verified on NMOS and vGPU setup,correct feature type is displayed in the message
over UI and in /var/log/messages as well.