Auto-reflect runtime changes in nvidia settings when the nvidia-gridd is restarted/stopped.
Issue and root cause:
When 'Manage License' page is created , a timer callback is registered to
query license state from nvidia-gridd with 1 sec time interval to
update the 'Manage License' page. When nvidia-gridd service is stopped,
there is no response against the Dbus request so the callback function exits with error
which destroys the timer. Hence 'Manage License' page is never updated.

Fix:
Do not destroy the timer which updates the 'Manage license' page even if 'nvidia-gridd' service
is stopped.
Upon failure to communicate with nvidia-gridd to fetch licensing state info
we are disabling all UI controls and notifying user about the issue via a message on the UI.

Testing Done:
Verified that if nvidia-gridd service is restarted then 'Manage License' page gets
refreshed automatically.