nvidia-settings: Don't query target relationships for nonexistent targets
Currently, nvidia-settings will ask the X driver about the relationships between
every type of target it knows about unconditionally.  When support for Canoas
targets in the X driver was removed (including the
NV_CTRL_BINARY_DATA_VCSCS_USED_BY_GPU attribute), this caused nvidia-settings to
trigger a BadMatch error and abort (on GTK2) or print an error about failing to
query relationships between targets (on GTK3).

This change updates nvidia-settings to not bother to query for relationships for
a target when no such targets exist in the system.  In addition to fixing the
BadMatch described above, it should reduce the startup latency due to fewer
server round-trips in the common case.