nvidia-settings: Ignore GPUs that can't be queried from NV-CONTROL
When AllowExternalGpus is not set to TRUE, the X server will ignore
surprise-removable GPUs.

Despite these GPUs being invisible to X and thus invisible to NV-CONTROL, they
are visible via NVML. nvidia-settings can use NVML to query for available GPUs,
and then query additional information from NV-CONTROL. Today, it assumes that a
GPU that is visible via NVML will also be visible via NV-CONTROL, but thanks to
Option "AllowExternalGpus", that is no longer true, resulting in breakage.

The same bug also occurs when using Option "ProbeAllGpus" "False". The effect is
the same as Option "AllowExternalGpus" "False", resulting in only a subset of
GPUs being visible via NV-CONTROL, compared to NVML.

As a short term solution, this change makes nvidia-settings do a mundane
NV-CONTROL query (NV_CTRL_DEPTH_30_ALLOWED) on any GPU target that it allocates,
silently ignoring GPUs that fail. This has the effect of making blacklisted
eGPUs (or other GPUs ignored due to Option "ProbeAllGpus" "False") invisible to
nvidia-settings, preventing further breakages.

NV_CTRL_DEPTH_30_ALLOWED was chosen because it has existed since 2007,
preventing breakage with currently supported drivers, and is unlikely to be
implemented in NVML. If it were to be implemented in NVML, the query could
succeed where we would expect it to fail, preventing the GPU from being properly
ignored.