Hide EGL Graphics Information instead of leaving a blank page if EGL queries fail.
If there is an issue querying graphics information for EGL on a system, we detect the error and hide all graphics information including OpenGL, Client/Server GLX info that may not be affected by the EGL issue. This leaves a blank page when only a portion of the code may have failed.

This change only omits the EGL info if EGL is the only problem.