Disable word-wrapping when output is not a terminal
Issue:

When redirecting nvidia-settings output to a file or a pipe (for simple
recording, or scripting purposes), we don't want the output to change
depending on the current terminal width (TIOCGWINSZ).

Fix/Feature:

This change disables word-wrapping of `nv_msg`/`nv_*_msg` functions when
the output stream is not a terminal (not `isatty`).

PRE-Merge Build Fix from Liam Middlebrook <lmiddlebrook@nvidia.com>:
    Add _GNU_SOURCE define for fileno